requirements.txt
google-cloud-storage==1.26.0
google-cloud-bigquery==1.24.0
numpy==1.17.0
pandas==1.0.1
python-dotenv==0.11.0
scikit-learn==0.21.3
tensorflow-hub==0.7.0
tensorflow<2.0
bert-tensorflow==1.0.1
pathlib 
regex
kneed
